requirements.txt
python-telegram-bot>=20.7
httpx>=0.27.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
